+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Custom variable `chronometrist-activity-indicator` to change how an active task is indicated.
 ### Fixed
 * `chronometrist-kv-accept` will not modify the file if there are no key-values.
 * Regression in `chronometrist-value-history-populate`